feat(guition): camera exposure + entropy-still follow-ups
- Fixed-exposure override: BOARD_CSI_AE_TARGET is written straight to
  V4L2_CID_EXPOSURE, so widen its range (235 -> 1200) and correct the help,
  which described a normalised setpoint it never was. Carry ae_target as
  uint16 -- a sensor whose exposure scale exceeds 255 (e.g. the OV02C10)
  otherwise truncates. A sensor with no on-sensor AE needs this fixed dial-back.

- Image-entropy still: a centred 720 square (was 960). The OV02C10 frame is
  only 728 tall, so 960 forced the still grab to upscale, which the PPA rejects.
  Re-pin esp-camera-pipeline to the rotation-aware grab.

- Document the cold-boot light-blue flash as a known cosmetic artifact of this
  panel so it is not re-attempted.

+++ src/board_pipeline_camera_csi.c
@@ -62,7 +62,7 @@ typedef struct {
     size_t frame_buf_size;
     uint16_t frame_w;
     uint16_t frame_h;
-    uint8_t ae_target;
+    uint16_t ae_target;             /* fixed V4L2_CID_EXPOSURE at start; 0 = sensor default */
     bool hmirror;
     bool vflip;
     cam_pipeline_frame_cb_t frame_cb;
